ref(kernel): new package, initramfs.py moves in
* kernel/ holds what shapes how the kernel runs on the target; models/kernel
  is its model the way models/swap and models/bootloader pair with theirs
