+++ installer/archinstoo/lib/kernel/initramfs.py
